### 2018-06-16 1f1b094
joss submission author string fix
