### 2018-06-24 #776 @arokem
Hey @nirum : would you be willing to review this paper for JOSS?
