### 2018-06-25 d58ea1c
joss submission: paper.md citation fixes
