### 2018-06-25 #792 @whedon
[REVIEW]: SeqPeelSort - A spike sorting algorithm for single sensillum recordings
**Submitting author:** @grg2rsr (<a href="http://orcid.org/0000-0002-9716-3513">Georg Raiser</a>)
**Repository:** <a href="https://github.com/grg2rsr/SeqPeelSort" target ="_blank">https://github.com/grg2rsr/SeqPeelSort</a>
**Version:** v1.0
