### 2018-06-25 #776 @arokem
Thanks @nirum! Head on over to #792 for the reviewer check-list. I'll close this one.
