### 2018-07-01 #792 @nirum
Added a couple of issues to the project repo. In general, looks great! I was able to install the code and run the example easily.
