### 2018-09-10 #792 @arokem
Hi @nirum : just to confirm -- does this mean that the paper is ready to be accepted from your point of view?

@grg2rsr : one request from my end: could you please add the DOI for the references in the paper?
