### 2018-09-10 fda51e6
DOIs added, some typos fixed
