+++ paper.bib
@@ -6,7 +6,8 @@
   number={7},
   pages={965--979},
   year={2004},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1016/j.cell.2004.05.012}
 }
 
 @article{lin2015re,
@@ -17,5 +18,6 @@
   number={10},
   pages={e0139675},
   year={2015},
-  publisher={Public Library of Science}
+  publisher={Public Library of Science},
+  doi = {10.1371/journal.pone.0139675}
 }
