### 2018-09-10 #792 @grg2rsr
DOIs are added
