### 2018-09-10 #792 @nirum
Yes, paper is ready to accept from my end!

> On Sep 10, 2018, at 5:44 AM, Georg Raiser <notifications@github.com> wrote:
>
> DOIs are added
>
