+++ README.md
@@ -8,7 +8,7 @@ _SeqPeelSort_ is a spike sorting algorithm for single sensillum recordings (SSR)
 
 Single sensillum recordings are a standard technique in insect neuroscience. In such electrophysiological recordings, a single electrode is inserted into a sensory sensillum, and the transepithelial potential is recorded. This technique is extensively (but not only) used in _Drosophila_ olfaction studies (see for example Hallem et al. (2004) or Lin et al. (2015) and citations therein).
 
-Since insect olfactory sensilla can house multiple sensory neurons, the action potentials of all those neurons are recorded with the single electrode simultaneously. The individual units usually display different spike shapes and amplitudes. Additionally, the amplitudes of some units is substantially decreased when the unit exhibits a high firing rate.
+Since insect olfactory sensilla can house multiple sensory neurons, the action potentials of all those neurons are recorded with the single electrode simultaneously. The individual units usually display different spike shapes and amplitudes. Additionally, the spike amplitude of some units is substantially decreased when the unit exhibits a high firing rate.
 
 As any spike sorting algorithm, _SeqPeelSorts_ goal is to detect spikes and based on their waveform, assign them to a unit. Different to most recording situations for which current spike sorting algorithms are designed, the recording configuration of SSR leads to specific constrains: The number of electrodes is only one, the number of units is known and low, but their spike shapes are quite variable. _SeqPeelSort_ is designed handle these specific constrains while being lightweight and simple.
 
