### 2018-09-17 #792 @grg2rsr
the error is still in the pdf, but not in the README as displayed in github. Attempting to generate pdf after some time delay in case something wasn't updated after first recompilation
