### 2018-09-23 #792 @arokem
@arfon: I believe this paper is ready to go!
