### 2018-09-23 #792 @arfon
@nirum many thanks for your review here and to @arokem for editing this submission ✨

@grg2rsr - your paper is now accepted into JOSS and your DOI is https://doi.org/10.21105/joss.00792 :zap: 🚀:boom:
